requirements.txt
beautifulsoup4==4.12.2
Flask==2.2.5
Flask-RESTful==0.3.10
gunicorn
requests==2.31.0
flask_httpauth
python-dotenv